add possibility to run `install.hs` from cabal
additionally provide mechanism for detecting if run from stack or cabal.
