+++ install/Setup.hs
@@ -1,0 +1,2 @@
+import Distribution.Simple
+main = defaultMain
