+++ install/hie-install.cabal
@@ -1,0 +1,28 @@
+name:                hie-install
+version:             0.8.0.0
+synopsis:            Install the haskell-ide-engine
+license:             BSD3
+author:              Many, TBD when we release
+maintainer:          samuel.pilz@posteo.net
+copyright:           2019
+build-type:          Simple
+cabal-version:       >=2.0
+
+library
+  hs-source-dirs:      .
+  exposed-modules:     Install
+  other-modules:       BuildSystem
+  build-depends:       base >= 4.9 && < 5
+                     , shake == 0.17.3
+                     , directory
+                     , extra
+                     , text
+  default-extensions: LambdaCase
+
+  if flag(run-from-stack)
+    cpp-options:       -DRUN_FROM_STACK
+
+flag run-from-stack
+  description:         Inform the application that it is run from stack
+  default:             False
+  manual:              True
