+++ shake.yaml
@@ -1,7 +1,11 @@
 # Used to provide a different environment for the shake build script
 resolver: lts-13.18 # GHC 8.6.4
 packages:
-- .
+- install
 
 nix:
   packages: [ zlib ]
+
+flags:
+  hie-install:
+    run-from-stack: true
